% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obtain.R
\name{obtain}
\alias{obtain}
\title{"obtain" data from various data sources}
\usage{
obtain(from = "csv", ...)
}
\arguments{
\item{from}{the type of data source from which the data will be obtained.
This should be one of "csv", "tsv", "excel", "rds", "mysql".}

\item{...}{any data source specific parameter}
}
\value{
a data frame containg the data
}
\description{
It is the first step in the \code{OSEMN} data science pipeline. It allows to read data from a chosen data source
}
\examples{
\dontrun{
#read data for a csv file
my_data <- obtain(from = "csv", file = "data.csv")

#read data for a csv file
my_data <- obtain(from = "csv", file = "data.csv")

#read data for a csv file
my_data <- obtain(from = "csv", file = "data.csv")

#read data for a csv file
my_data <- obtain(from = "csv", file = "data.csv")
}

}
